Move implementation details into compose
