+++ prototype/luamake/Makefile
@@ -4,4 +4,5 @@ hello:
 include luamake.mk
 
 $(lua require,compose)
+$(lua require,mutation)
 $(lua require,example)
